fix(p15): apply QA review remediation — spec, findings, shim comment, cast, TLS test
R-1  spec.md: corrected suppression description from "all 8 FT8 tone bins"
     to "exact decoded tone bin ±1 nearest neighbours (Hann-window sidelobes)"
     per design.md §Decision 2; added AC-IS-1 through AC-IS-5 acceptance
     criteria with measured results and Captain's Option A decision recorded.

R-2  Ft8LibInteropTests.cs: new Ft8LibInterop TLS regression test —
     GetLastPassCounts returns [0,0] after DecodeAll on a silent PCM buffer.

R-3  ft8_shim.c: removed spurious (uint16_t) cast in dedup slot calculation;
     corrected header comment to 69.1% / 613 (was 69.0% / 612).
     libft8.dll rebuilt (win-x64); libft8.version.txt updated.

R2-1 findings.md: replaced stale "not implementing iterative subtraction /
     deferred to a future change" text with accurate post-p15 description:
     implemented in p15 via ±1-bin narrow suppression; remaining ~31% gap
     is the spectrogram-domain ceiling; PCM-domain approach is future work.

QA artefacts: qa-review.md (APPROVED ✅), dev-action-items.md committed.

+++ src/OpenWSFZ.Ft8/Native/ft8_shim.c
@@ -22,7 +22,7 @@
  *   domain suppression is the most accurate approach available without a
  *   dedicated sub-Hz carrier-frequency estimator.
  *
- *   Measured result on the 42-cycle corpus: 69.0% recovery (612/887).
+ *   Measured result on the 42-cycle corpus: 69.1% recovery (613/887).
  *   Baseline (single pass): 66.6% (591/887).
  *
  * Build: see BUILD.md.  encode.c must be compiled and linked.
@@ -269,7 +269,7 @@ int ft8_decode_all(
                 continue;
 
             /* Cross-pass dedup */
-            int  slot = (int)(msg.hash % (uint16_t)K_MAX_DECODED);
+            int  slot = (int)(msg.hash % K_MAX_DECODED);
             bool dup  = false, found = false;
             int  walk = slot;
             for (int s = 0; s < K_MAX_DECODED; ++s) {
